requirements.txt
Appium-Python-Client
Flask
selenium~=4.12.0